requirements.txt
# =============================================================================
# DEPENDÊNCIAS DO PROJETO SERVIMED - DESAFIO TÉCNICO
# =============================================================================
# Este arquivo define todas as dependências necessárias para o projeto
# Versões específicas garantem estabilidade e reprodutibilidade
# =============================================================================

# =============================================================================
# DEPENDÊNCIAS PRINCIPAIS - CORE DO PROJETO
# =============================================================================

# Framework de web scraping
Scrapy==2.11.0

# Gerenciamento de variáveis de ambiente
python-dotenv==1.0.0

# Requisições HTTP para APIs
requests==2.31.0

# =============================================================================
# DEPENDÊNCIAS FASE 2 - SISTEMA ASSÍNCRONO
# =============================================================================

# Framework para workers assíncronos
celery==5.3.4

# Framework web moderno para API
fastapi==0.104.1

# Servidor ASGI para FastAPI
uvicorn[standard]==0.24.0

# Validação de dados para FastAPI
pydantic==2.5.0

# =============================================================================
# DEPENDÊNCIAS DE AUTENTICAÇÃO E SEGURANÇA
# =============================================================================

# Criptografia e segurança
cryptography==41.0.7

# Validação de dados
pydantic==2.5.0

# =============================================================================
# DEPENDÊNCIAS DE PROCESSAMENTO DE DADOS
# =============================================================================

# Manipulação de dados estruturados
pandas==2.1.4

# Processamento de JSON avançado
jsonschema==4.20.0

# =============================================================================
# DEPENDÊNCIAS DE LOGGING E MONITORAMENTO
# =============================================================================

# Logging estruturado
structlog==23.2.0

# Métricas e monitoramento
prometheus-client==0.19.0

# =============================================================================
# DEPENDÊNCIAS DE CACHE E PERFORMANCE
# =============================================================================

# Cache em memória
redis==5.0.1

# Serialização rápida
orjson==3.9.10

# =============================================================================
# DEPENDÊNCIAS DE TESTES
# =============================================================================

# Framework de testes
pytest==7.4.3

# Mocks para testes
pytest-mock==3.12.0

# Cobertura de código
pytest-cov==4.1.0

# Testes de performance
pytest-benchmark==4.0.0

# =============================================================================
# DEPENDÊNCIAS DE QUALIDADE DE CÓDIGO
# =============================================================================

# Formatação de código
black==23.12.0

# Linting e análise estática
flake8==6.1.0

# Verificação de tipos
mypy==1.8.0

# Verificação de segurança
bandit==1.7.5

# =============================================================================
# DEPENDÊNCIAS DE DESENVOLVIMENTO
# =============================================================================

# Pre-commit hooks
pre-commit==3.6.0

# Geração de documentação
sphinx==7.2.6

# =============================================================================
# DEPENDÊNCIAS DE DEPLOY E INFRAESTRUTURA
# =============================================================================

# Containerização
docker==6.1.3

# Orquestração
docker-compose==1.29.2

# =============================================================================
# DEPENDÊNCIAS OPCIONAIS - DESCOMENTAR CONFORME NECESSIDADE
# =============================================================================

# # Banco de dados PostgreSQL
# psycopg2-binary==2.9.9

# # Banco de dados MongoDB
# pymongo==4.6.0

# # Filas assíncronas
# celery==5.3.4

# # Cache distribuído
# memcached==1.59

# # Monitoramento avançado
# sentry-sdk==1.38.0

